Move client from apps into client dir
